Modularized page.tsx
created a signup component
